requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
redis
openai
